+++ src/pages/CommitteePage/InCommitteeGrid.tsx
@@ -2,23 +2,12 @@ import { PageInfo } from "@digitalaidseattle/core";
 import { Link as MuiLink } from "@mui/material";
 import { DataGrid, GridColDef, GridRenderCellParams, useGridApiRef } from "@mui/x-data-grid";
 import { useEffect, useMemo, useState } from "react";
-import { Link as RouterLink } from "react-router-dom";
-import { Bill, LegislationInfo } from "../../api/bill";
+import { NavLink } from "react-router-dom";
 import { BillService } from "../../api/billService";
 import { Committee } from "../../api/committee";
+import { CommitteeBillRow, mapCommitteeBillRow } from "../../utils/committees";
 
 const PAGE_SIZE = 25;
-
-type CommitteeBillRow = {
-  id: string;
-  bill: string;
-  originalSponsor: string;
-  title: string;
-  status: string;
-  history: string;
-  billPageId?: string;
-  billSearchUrl: string;
-};
 
 const InCommitteeGrid = (props: { committee: Committee, search?: string, refreshKey?: number }) => {
   const apiRef = useGridApiRef();
@@ -61,9 +50,9 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
       renderCell: (params: GridRenderCellParams<CommitteeBillRow>) => {
         if (params.row.billPageId) {
           return (
-            <MuiLink component={RouterLink} to={`/bill/${params.row.billPageId}`} underline="always" color="text.primary">
+            <NavLink to={`/bill/${params.row.billPageId}`}>
               {params.row.bill}
-            </MuiLink>
+            </NavLink>
           );
         }
 
@@ -80,11 +69,7 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
       minWidth: 180,
       flex: 0.9,
       type: "string",
-      renderCell: (params: GridRenderCellParams<CommitteeBillRow>) => (
-        <MuiLink component="span" underline="always" color="text.primary">
-          {params.row.originalSponsor}
-        </MuiLink>
-      )
+      renderCell: (params: GridRenderCellParams<CommitteeBillRow>) => params.row.originalSponsor
     },
     {
       field: "title",
@@ -116,9 +101,9 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
       renderCell: (params: GridRenderCellParams<CommitteeBillRow>) => {
         if (params.row.billPageId) {
           return (
-            <MuiLink component={RouterLink} to={`/bill/${params.row.billPageId}`} underline="always">
+            <NavLink to={`/bill/${params.row.billPageId}`}>
               Link to Bill page
-            </MuiLink>
+            </NavLink>
           );
         }
 
@@ -147,7 +132,7 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
     BillService.getInstance()
       .getAll()
       .then((bills) => {
-        const rows = inCommittee.map((legislation) => toCommitteeBillRow(legislation, bills));
+        const rows = inCommittee.map((legislation) => mapCommitteeBillRow(legislation, bills));
         setPageInfo({
           rows,
           totalRowCount: rows.length,
@@ -155,79 +140,12 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
       })
       .catch((error) => {
         console.error('Error loading committee bill details:', error);
-        const rows = inCommittee.map((legislation) => toCommitteeBillRow(legislation, []));
+        const rows = inCommittee.map((legislation) => mapCommitteeBillRow(legislation, []));
         setPageInfo({
           rows,
           totalRowCount: rows.length,
         });
       });
-  }
-
-  function toCommitteeBillRow(
-    legislation: LegislationInfo,
-    bills: Bill[]
-  ): CommitteeBillRow {
-    const bill = findBill(legislation, bills);
-
-    return {
-      id: legislation.BillId,
-      bill: bill?.BillId ?? legislation.BillId,
-      originalSponsor: formatSponsor(bill),
-      title: bill?.LegalTitle ?? bill?.LongDescription ?? bill?.ShortDescription ?? "",
-      status: bill?.CurrentStatus?.Status ?? "",
-      history: formatHistory(bill),
-      billPageId: bill?.BillId,
-      billSearchUrl: getBillSearchUrl(bill, legislation)
-    };
-  }
-
-  function findBill(legislation: LegislationInfo, bills: Bill[]) {
-    const normalizedBillId = normalizeBillId(legislation.BillId);
-    return bills.find((bill) =>
-      normalizeBillId(bill.BillId) === normalizedBillId ||
-      String(bill.BillNumber) === String(legislation.BillNumber)
-    );
-  }
-
-  function normalizeBillId(value: string | number | undefined) {
-    return String(value ?? "").replace(/\s+/g, "").toUpperCase();
-  }
-
-  function formatSponsor(bill?: Bill) {
-    const sponsor = bill?.Sponsors?.[0];
-    if (sponsor) {
-      return sponsor.Name || `${sponsor.FirstName ?? ""} ${sponsor.LastName ?? ""}`.trim();
-    }
-    return bill?.Sponsor ?? "";
-  }
-
-  function formatHistory(bill?: Bill) {
-    if (!bill?.CurrentStatus) {
-      return "";
-    }
-    const date = formatDate(bill.CurrentStatus.ActionDate);
-    return [date, bill.CurrentStatus.HistoryLine].filter(Boolean).join(" ");
-  }
-
-  function formatDate(raw?: string) {
-    if (!raw) {
-      return "";
-    }
-    const date = new Date(raw);
-    if (Number.isNaN(date.getTime())) {
-      return raw;
-    }
-    return new Intl.DateTimeFormat("en-US", {
-      month: "short",
-      day: "numeric"
-    }).format(date);
-  }
-
-  function getBillSearchUrl(bill: Bill | undefined, legislation: LegislationInfo) {
-    const biennium = bill?.Biennium ?? legislation.Biennium;
-    const year = biennium?.split("-")[0] ?? "";
-    const billNumber = bill?.BillNumber ?? legislation.BillNumber;
-    return `https://app.leg.wa.gov/billsummary/?BillNumber=${billNumber}&Year=${year}`;
   }
 
   return (
@@ -242,17 +160,6 @@ const InCommitteeGrid = (props: { committee: Committee, search?: string, refresh
       pageSizeOptions={[10, 25, 50, 100]}
       getRowHeight={() => "auto"}
       disableRowSelectionOnClick
-      sx={{
-        border: 0,
-        "& .MuiDataGrid-columnHeaderTitle": { fontWeight: 700 },
-        "& .MuiDataGrid-cell": {
-          alignItems: "center",
-          minHeight: "96px",
-          py: 2,
-          whiteSpace: "normal",
-          lineHeight: 1.25
-        }
-      }}
     />
   );
 };
