+++ src/pages/CommitteesPage.tsx
@@ -4,16 +4,18 @@
  *  @copyright 2026 Digital Aid Seattle
  *
  */
-import { ExpandAltOutlined, ExportOutlined, ReloadOutlined, SearchOutlined } from "@ant-design/icons";
+import { ExpandAltOutlined, ExportOutlined, HomeOutlined, ReloadOutlined } from "@ant-design/icons";
 import { FilterItem, LoadingContext, PageInfo, QueryModel, useNotifications } from "@digitalaidseattle/core";
-import { Box, Card, CardContent, IconButton, InputAdornment, Link as MuiLink, TextField, Toolbar, Tooltip, Typography } from '@mui/material';
+import { Breadcrumbs, Card, CardHeader, IconButton, Toolbar, Tooltip, Typography } from '@mui/material';
 import { DataGrid, GridColDef, GridFilterModel, GridRenderCellParams, GridSortModel, useGridApiRef } from "@mui/x-data-grid";
 import { useContext, useEffect, useMemo, useState } from 'react';
-import { Link as RouterLink, useNavigate } from "react-router-dom";
+import { NavLink, useNavigate } from "react-router-dom";
 import { Committee } from '../api/committee';
 import { LegislatureService } from '../api/legislatureService';
 import { CHAMBER_TYPE, ChamberButtonGroup } from "../components/ChamberButtonGroup";
 import { LoadingOverlay } from "../components/LoadingOverlay";
+import { SearchField } from "../components/SearchField";
+import { formatCommitteeName, getLeadershipName } from "../utils/committees";
 
 const PAGE_SIZE = 25;
 const COMMITTEE_SEARCH_URL = "https://leg.wa.gov/about-the-legislature/committees/";
@@ -39,15 +41,9 @@ const CommitteesPage = () => {
       flex: 1.5,
       type: "string",
       renderCell: (params: GridRenderCellParams<Committee>) => (
-        <MuiLink
-          component={RouterLink}
-          to={`/committee/${params.row.Id}`}
-          underline="always"
-          color="text.primary"
-          sx={{ whiteSpace: "normal", lineHeight: 1.25 }}
-        >
+        <NavLink to={`/committee/${params.row.Id}`}>
           {formatCommitteeName(params.row)}
-        </MuiLink>
+        </NavLink>
       )
     },
     {
@@ -56,7 +52,7 @@ const CommitteesPage = () => {
       minWidth: 160,
       flex: 0.8,
       valueGetter: (_value, row) => getLeadershipName(row, "chair"),
-      renderCell: (params: GridRenderCellParams<Committee>) => renderName(getLeadershipName(params.row, "chair")),
+      renderCell: (params: GridRenderCellParams<Committee>) => getLeadershipName(params.row, "chair"),
       type: "string",
       sortable: false,
       filterable: false
@@ -67,7 +63,7 @@ const CommitteesPage = () => {
       minWidth: 170,
       flex: 0.85,
       valueGetter: (_value, row) => getLeadershipName(row, "vice"),
-      renderCell: (params: GridRenderCellParams<Committee>) => renderName(getLeadershipName(params.row, "vice")),
+      renderCell: (params: GridRenderCellParams<Committee>) => getLeadershipName(params.row, "vice"),
       type: "string",
       sortable: false,
       filterable: false
@@ -78,7 +74,7 @@ const CommitteesPage = () => {
       minWidth: 180,
       flex: 0.9,
       valueGetter: (_value, row) => getLeadershipName(row, "ranking"),
-      renderCell: (params: GridRenderCellParams<Committee>) => renderName(getLeadershipName(params.row, "ranking")),
+      renderCell: (params: GridRenderCellParams<Committee>) => getLeadershipName(params.row, "ranking"),
       type: "string",
       sortable: false,
       filterable: false
@@ -89,7 +85,7 @@ const CommitteesPage = () => {
       minWidth: 180,
       flex: 0.9,
       valueGetter: (_value, row) => getLeadershipName(row, "majority"),
-      renderCell: (params: GridRenderCellParams<Committee>) => renderName(getLeadershipName(params.row, "majority")),
+      renderCell: (params: GridRenderCellParams<Committee>) => getLeadershipName(params.row, "majority"),
       type: "string",
       sortable: false,
       filterable: false
@@ -185,120 +181,40 @@ const CommitteesPage = () => {
 
   function CustomToolbar() {
     return (
-      <Toolbar
-        disableGutters
-        sx={{
-          alignItems: { xs: "stretch", md: "center" },
-          gap: 1.5,
-          justifyContent: "space-between",
-          flexWrap: "wrap",
-          mb: 0.25
-        }}
-      >
+      <Toolbar>
         <ChamberButtonGroup chamber={chamber} onChange={handleChamberChange} />
-        <Box sx={{ display: "flex", alignItems: "center", gap: 1, ml: { md: "auto" } }}>
-          <TextField
-            value={search}
-            onChange={(event) => setSearch(event.target.value)}
-            size="small"
-            placeholder="Search"
-            aria-label="Search committees"
-            sx={{ width: { xs: "100%", sm: 240 } }}
-            slotProps={{
-              input: {
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <SearchOutlined />
-                  </InputAdornment>
-                )
-              }
-            }}
-          />
-          <Tooltip title="Open committee directory">
-            <IconButton
-              color="primary"
-              onClick={() => window.open(COMMITTEE_SEARCH_URL, "_blank", "noopener")}
-              aria-label="Open committee directory"
-            >
-              <ExpandAltOutlined />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title="Export">
-            <IconButton color="primary" size="small" onClick={exportData} aria-label="Export committees">
-              <ExportOutlined />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title="Refresh">
-            <IconButton color="primary" size="small" onClick={refresh} aria-label="Refresh committees">
-              <ReloadOutlined />
-            </IconButton>
-          </Tooltip>
-        </Box>
+        <SearchField value={search} onChange={(value) => setSearch(value)} />
+        <Tooltip title="Open committee directory">
+          <IconButton
+            color="primary"
+            onClick={() => window.open(COMMITTEE_SEARCH_URL, "_blank", "noopener")}
+            aria-label="Open committee directory"
+          >
+            <ExpandAltOutlined />
+          </IconButton>
+        </Tooltip>
+        <Tooltip title="Export">
+          <IconButton color="primary" size="small" onClick={exportData} aria-label="Export committees">
+            <ExportOutlined />
+          </IconButton>
+        </Tooltip>
+        <Tooltip title="Refresh">
+          <IconButton color="primary" size="small" onClick={refresh} aria-label="Refresh committees">
+            <ReloadOutlined />
+          </IconButton>
+        </Tooltip>
       </Toolbar>
     );
   }
 
-  function getLeadershipName(committee: Committee, role: string) {
-    const leadership = committee.Leadership ?? [];
-    const found = leadership.find((leader) => {
-      const normalizedRole = leader.role.toLowerCase();
-
-      if (role === "chair") {
-        return normalizedRole.includes("chair") &&
-          !normalizedRole.includes("vice") &&
-          !normalizedRole.includes("minority") &&
-          !normalizedRole.includes("majority");
-      }
-
-      if (role === "vice") {
-        return normalizedRole.includes("vice");
-      }
-
-      if (role === "ranking") {
-        return normalizedRole.includes("ranking") &&
-          !normalizedRole.includes("assistant");
-      }
-
-      if (role === "majority") {
-        return normalizedRole.includes("majority");
-      }
-
-      return false;
-    });
-    return formatPersonName(found?.name ?? "");
-  }
-
-  function formatCommitteeName(committee: Committee) {
-    return committee.LongName || committee.Name;
-  }
-
-  function renderName(name: string) {
-    if (!name) {
-      return "";
-    }
-
-    return (
-      <MuiLink component="span" underline="always" color="text.primary">
-        {name}
-      </MuiLink>
-    );
-  }
-
-  function formatPersonName(name: string) {
-    const [lastName, firstName] = name.split(",").map((part) => part.trim());
-    if (firstName && lastName) {
-      return `${firstName} ${lastName}`;
-    }
-    return name;
-  }
-
   return (<>
     <LoadingOverlay loading={loading} />
-    <Card sx={{ mx: { xs: 1, md: 3 }, my: { xs: 2, md: 4 } }}>
-      <CardContent>
-        <Typography component="h1" variant="h3" sx={{ fontSize: { xs: 28, md: 36 }, fontWeight: 700, mb: 2.5 }}>
-          Committees
-        </Typography>
+    <Breadcrumbs aria-label="breadcrumb">
+      <NavLink to="/" ><IconButton size="medium"><HomeOutlined /></IconButton></NavLink>
+      <Typography color="text.primary">Committees</Typography>
+    </Breadcrumbs>
+    <Card>
+      <CardHeader title="Committees" />
         <DataGrid
           apiRef={apiRef}
           autoHeight
@@ -320,15 +236,7 @@ const CommitteesPage = () => {
           showToolbar={true}
           slots={{ toolbar: CustomToolbar }}
           disableRowSelectionOnClick
-          sx={{
-            border: 0,
-            "& .MuiDataGrid-columnHeaders": { borderTop: "1px solid", borderColor: "divider" },
-            "& .MuiDataGrid-columnHeaderTitle": { fontWeight: 700 },
-            "& .MuiDataGrid-row": { cursor: "pointer" },
-            "& .MuiDataGrid-cell": { py: 2, alignItems: "center" }
-          }}
         />
-      </CardContent>
     </Card>
   </>);
 };
